# docs/openapi/climate.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CHART API",
    "description": "\nThe single CHART application API for authentication, geography-scoped planning,\nclimate data preparation, model predictions, setup, users, workspaces, hazards,\nand public solutions.\n\nProtected routes use a Keycloak bearer token and enforce both role and geography\nscope. Prediction submission is durable: HTTP 200 means a completed request was\navailable, while HTTP 202 means Dagster has queued background work. Swagger is\navailable at `/docs` and ReDoc at `/redoc`.\n",
    "version": "0.2.0"
  },
  "paths": {
...
    "/model-catalog": {
      "get": {
        "tags": [
          "model-catalog"
        ],
        "summary": "List active model releases known to this installation",
        "description": "Returns each currently active ModelRelease with its climate hazard, health outcome, and covered admin_units so the planning UI can render hazard and outcome pickers straight from the registry instead of a client-side lookup table. Records reflect the ActiveModelAssignment state at the moment of the call.",
        "operationId": "list_catalog_model_catalog_get",
        "parameters": [
          {
            "name": "geography_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Geography Id"
            }
          },
          {
            "name": "include_descendants",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Descendants"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The active model releases with taxonomy fields and the admin_units each release is currently assigned to.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/model-releases/{release_id}/reload": {
      "post": {
        "tags": [
          "model-releases"
        ],
        "summary": "Re-verify and re-warm one release into the R runtime",
        "description": "Finds the manifest whose id matches the URL parameter, re-parses its ModelReleaseSpec, and calls prepare_model_release to re-POST /models/load to the R runtime. Does not touch the database or reactivate the release. Useful after an R container restart or after a manifest edit that swaps an artifact's SHA-256.",
        "operationId": "reload_release_model_releases__release_id__reload_post",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The release id was found on disk and successfully warmed into the runtime; status is 'loaded'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogEntry": {
        "properties": {
          "climate_hazard": {
            "type": "string",
            "title": "Climate Hazard"
          },
          "climate_hazard_label": {
            "type": "string",
            "title": "Climate Hazard Label"
          },
          "health_domain": {
            "type": "string",
            "title": "Health Domain"
          },
          "health_domain_label": {
            "type": "string",
            "title": "Health Domain Label"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "outcome_label": {
            "type": "string",
            "title": "Outcome Label"
          },
          "dashboard_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashboard Title"
          },
          "population_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population Label"
          },
          "model_scope_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Scope Label"
          },
          "effect_measure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect Measure"
          },
          "batch_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Status"
          },
          "visualization_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization Type"
          },
          "visualization_figure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization Figure"
          },
          "visualization_context_figure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization Context Figure"
          },
          "risk_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Description"
          },
          "release_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Release Ids"
          }
        },
        "type": "object",
        "required": [
          "climate_hazard",
          "climate_hazard_label",
          "health_domain",
          "health_domain_label",
          "outcome",
          "outcome_label",
          "release_ids"
        ],
        "title": "CatalogEntry"
      },
      "CatalogResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CatalogResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReloadResponse": {
        "properties": {
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "status"
        ],
        "title": "ReloadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Keycloak access token",
        "scheme": "bearer"
      },
      "chart-internal-token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}